Add files via upload
This version has additions to the method (now called fillArray) that are meant to convert the char at a given point in the string entry to an integer. This integer has 65 subtracted from it which should yield the (ascii value of that char) - 0. this value is the index for the array "tally" which will hold a string of identical chars with one being added each time a char is analyzed from the string. Right now this array does not work and has many errors. I currently am trying to make the array in a parent class to be inherited by the main header file method class. I would like this to work because it would be neater than passing it back and forth between main and header file and it would help me learn more about inheritance in general.