Add files via upload
This update is functional in that it can take a string, put it's ASCII values into an array of integers and then re interpret that array into chars and print the chars to console in alphabetical order with correct quantities of each. It can also print the size of the string. It does not have a catch for invalid characters (!#$%3567 " ect.) but that will be the next step. I Also abandoned the inheritance goal. It was going to be too much to tackle on my first c++ project. I worked on it for a while and was stuck. Next I want to add a input for a second string and be able to add them together. That will be the end goal.